refactor(agent): split agent domain into definition, registry, discovery sub-domains
