+++ packages/agent/src/agent/definition/capabilities.ts
