+++ packages/agent/src/agent/definition/index.ts
@@ -1,0 +1,24 @@
+export {
+  AgentProfileSchema,
+  type AgentProfile,
+  AgentIdentitySchema,
+  type AgentIdentity,
+  AgentStatusSchema,
+  type AgentStatus,
+  createAgentIdentity,
+} from "./profile";
+
+export {
+  AgentCapabilitiesSchema,
+  type AgentCapabilities,
+  DataScopeSchema,
+  type DataScope,
+  PolicySpecSchema,
+  type PolicySpec,
+} from "./capabilities";
+
+export {
+  AgentRuntimeSchema,
+  type AgentRuntime,
+  createAgentRuntime,
+} from "./runtime";
