+++ packages/agent/src/agent/discovery/frontmatter.ts
