+++ packages/agent/src/agent/discovery/index.ts
@@ -1,0 +1,7 @@
+export { parseFrontmatter } from "./frontmatter";
+
+export {
+  AgentDiscovery,
+  type AgentLoadResult,
+  type AgentDiscoveryOptions,
+} from "./discovery";
