+++ packages/agent/src/agent/index.ts
@@ -1,54 +1,6 @@
-export {
-  AgentProfileSchema,
-  type AgentProfile,
-  AgentIdentitySchema,
-  type AgentIdentity,
-  AgentStatusSchema,
-  type AgentStatus,
-  createAgentIdentity,
-} from "./profile";
+// Agent domain — identity, registry, discovery, messaging
 
-export {
-  AgentRegistry,
-  InMemoryAgentRegistryStore,
-  type AgentRegistryStore,
-} from "./profile-store";
-
-export {
-  AgentCapabilitiesSchema,
-  type AgentCapabilities,
-  DataScopeSchema,
-  type DataScope,
-  PolicySpecSchema,
-  type PolicySpec,
-} from "./capabilities";
-
-export {
-  AgentRuntimeSchema,
-  type AgentRuntime,
-  createAgentRuntime,
-} from "./runtime";
-
-export { FileAgentRegistryStore } from "./file-registry-storage";
-
-export {
-  AgentMessenger,
-  type MessageEnvelope,
-  type DeliveryOptions,
-  MessagingError,
-  type AllowPattern,
-} from "./communication";
-
-export {
-  BuiltinAgentRegistry,
-  AgentDefinitionSchema,
-  type AgentDefinition,
-} from "./registry";
-
-export { parseFrontmatter } from "./frontmatter";
-
-export {
-  AgentDiscovery,
-  type AgentLoadResult,
-  type AgentDiscoveryOptions,
-} from "./discovery";
+export * from "./definition";
+export * from "./registry";
+export * from "./discovery";
+export * from "./communication";
