+++ packages/agent/src/agent/registry/index.ts
@@ -1,0 +1,13 @@
+export {
+  BuiltinAgentRegistry,
+  AgentDefinitionSchema,
+  type AgentDefinition,
+} from "./registry";
+
+export {
+  AgentRegistry,
+  InMemoryAgentRegistryStore,
+  type AgentRegistryStore,
+} from "./profile-store";
+
+export { FileAgentRegistryStore } from "./file-registry-storage";
