+++ packages/agent/src/agent/registry/registry.ts
