+++ packages/agent/test/agent/file-registry-storage.test.ts
@@ -2,9 +2,9 @@ import { describe, expect, test, beforeEach, afterEach } from "bun:test";
 import { mkdtempSync, rmSync, existsSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-import { FileAgentRegistryStore } from "../../src/agent/file-registry-storage";
-import type { AgentProfile } from "../../src/agent/profile";
-import { AgentRegistry } from "../../src/agent/profile-store";
+import { FileAgentRegistryStore } from "../../src/agent/registry/file-registry-storage";
+import type { AgentProfile } from "../../src/agent/definition/profile";
+import { AgentRegistry } from "../../src/agent/registry/profile-store";
 
 function makeProfile(id: string, name?: string): AgentProfile {
   return { id, name: name ?? `Agent ${id}` };
